Fix implementations, add ID log and test data
Fixes, a lot of fixes. This time though, an ID class has been added. It manages the log of a unique ID for every created item.

Former system based every item's id on the index of its column array. Made everything more complicated than it should be.

There are still fixes to be done, but for now, everything mostly works as intended.